Separate out all the tests
